separate server and kernel python environments
This is a better and cleaner architecture which makes it easy to update
the kernels and libraries used by end-users in isolation from the
Jupyter server.

Basically: use whichever Python is most convenient for running the
server, and add kernels using conda. See guide in README.md for further
details.