Use pyenv instead of conda
I don't use conda otherwise and it's just too much of a hassle to have
to dig around the documentation each time I need to do something with
it. Also, it turns out it's much simpler NOT to have separate Python
environments for running Jupyter/JupyterHub and for the kernels. It
saves a lot of headscratching re: where to install Jupyter-related
libraries (because some are actually also needed by the kernel Python).